requirements.txt
fastapi
mypy
pydantic
python-dotenv
PyYAML
requests
RPi.GPIO
uvicorn
